# src/main/resources/api.yml
openapi: 3.0.0
info:
  contact:
    email: randakar@gmail.com
  description: Authenticare banking cares for your authentic experience!
  license:
    name: The GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  termsOfService: https://authenticare-bank.com/terms
  title: AuthentiCare Bank
  version: 0.0.1
externalDocs:
  description: Find this on github!
  url: https://github.com/randakar/AuthentiCareBankApi
servers:
  - url: http://localhost:8080/
  - url: /
security:
  - BasicAuth: []
tags:
  - description: Operations on accounts
    name: account
  - description: Operations on customers
    name: customer
  - description: Money transfer
    name: transaction
paths:

  /customer/{id}:
    get:
      operationId: getCustomerById
      parameters:
        - description: 'The id of the customer that needs to be fetched. '
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerViewModel'
          description: Customer view model retrieved
        404:
          content: {}
          description: Customer not found
      summary: Get customer by id
      tags:
        - customer
      x-accepts: application/json

  /customer/{id}/overview:
    get:
      operationId: getCustomerOverview
      parameters:
        - description: 'The id of the customer to be fetched for'
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/CustomerOverviewModel'
          description: successful operation
        404:
          content: {}
          description: Customer not found
      summary: Get customer by id
      tags:
        - customer
      x-accepts: application/json

  /customer/{id}/accounts:
    get:
      operationId: getCustomerAccounts
      parameters:
        - description: 'The id of the customer that accounts need to be fetched for'
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountViewModel'
          description: successful operation
        404:
          content: {}
          description: Customer not found
      summary: Get customer by id
      tags:
        - customer
      x-accepts: application/json

  /customer:
    get:
      operationId: getCurrentCustomer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerViewModel'
          description: successful operation
        404:
          content: { }
          description: Customer not found
      summary: Get my own customer details
      tags:
        - customer
      x-accepts: application/json

    post:
      operationId: createCustomer
      summary: Create a new customer.
      description: Create a new customer.
      requestBody:
        description: Customer
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerModel'
      responses:
        200:
          description: Customer created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerViewModel'
        409:
          description: Customer already exists.
      x-accepts: application/json
      tags:
        - customer


  /account/{id}:
    get:
      operationId: getAccountById
      parameters:
        - description: 'The id of the account that needs to be fetched. '
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountViewModel'
          description: Account view model retrieved
        404:
          content: {}
          description: Account not found
      summary: Get account by id
      tags:
        - account
      x-accepts: application/json

  /account/{id}/overview:
    get:
      operationId: getAccountOvervieuwById
      parameters:
        - description: 'The id of the account that needs to be fetched. '
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewModel'
          description: Account view model retrieved
        404:
          content: {}
          description: Account not found
      summary: Get account by id
      tags:
        - account
      x-accepts: application/json

  /account:
    get:
      operationId: getMyAccounts
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountViewModel'
          description: successful operation
        404:
          content: { }
          description: Customer not found
      summary: Get my own account details
      tags:
        - account
      x-accepts: application/json

    post:
      operationId: createAccount
      summary: Create a new account.
      description: Create a new account.
      requestBody:
        description: Account
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountModel'
      responses:
        200:
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountViewModel'
        409:
          description: Customer already exists.
      x-accepts: application/json
      tags:
        - account

  /account/{id}/transactions:
    summary: Transfer money
    get:
      operationId: getTransactions
      parameters:
        - description: 'The id of the account for which all transactions need to be shown'
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionViewModel'
          description: Transactions retrieved
        404:
          content: {}
          description: Account not found
      summary: Get transactions for account by account id
      tags:
        - transaction
      x-accepts: application/json

    post:
      operationId: transferAmount
      description: Transfer money
      parameters:
        - description: 'The id of the account for which the transfer applies'
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        description: Money transfer
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferModel'
      responses:
        201:
          description: Transaction successful
      x-accepts: application/json
      summary: Transfer money between accounts
      tags:
        - transaction

  /account/{id}/transactions/withdraw:
    post:
      operationId: withdrawAmount
      description: Withdraw cash
      parameters:
        - description: 'The id of the account for which the withdrawal applies'
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: integer
          style: simple

      requestBody:
        description: Money transfer to ATM
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithDrawModel'
      responses:
        201:
          description: Transaction successful
      summary: Withdraw money from account
      x-accepts: application/json
      tags:
        - transaction




components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

  schemas:

    currencyCode:
      type: string
      description: 3 letter currency code as defined by ISO-4217
      pattern: "[A-Z]{3}"
      format: iso-4217
      example: EUR

    iban:
      type: string
      pattern: '[A-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}'
      example: 'NL20INGB0001234567'

    transactionType:
      type: string
      pattern: (withdrawal|transfer)


    MoneyModel:
      type: object
      properties:
        amount:
          type: number
          description: Amount expressed as a decimal number of major currency units
          format: decimal
          minimum: 0
          example: 99.95
        currency:
          $ref: '#/components/schemas/currencyCode'
      required:
        - amount
        - currency

    CustomerModel:
      example:
        userName: Fooman
        email: email@rabobank.nl
        firstName: Fooman
        lastName: Barname
      properties:
        userName:
          type: string
          pattern: '[A-Z]{3,}'
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - userName
      type: object

    CustomerViewModel:
      allOf: # Complete view
        - $ref: '#/components/schemas/CustomerModel'
      example:
        id: 0
      properties:
        id:
          type: integer
          format: int64
      required:
        - id
        - userName

    CustomerOverviewModel:
      allOf: # Complete view
        - $ref: '#/components/schemas/CustomerViewModel'
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountViewModel'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardModel'

    AccountModel:
      example:
        iban: 'NL20INGB0001234567'
        owner: 0
        name: "DeBetaalRekening"
        description: 'Mijn Eerste Rekening!'
      properties:
        iban:
          $ref: '#/components/schemas/iban'
        name:
          type: string
        description:
          type: string
        owners:
          type: array
          items:
            type: integer
            format: int64
      required:
        - iban
        - owners

    AccountViewModel:
      allOf: # Complete view
        - $ref: '#/components/schemas/AccountModel'
      example:
        id: 0
      properties:
        id:
          type: integer
          format: int64
        balance:
          $ref: '#/components/schemas/MoneyModel'
      required:
        - id

    AccountOverviewModel:
      allOf: # Complete view
        - $ref: '#/components/schemas/AccountViewModel'
      properties:
        owners:
          type: array
          items:
            $ref: '#/components/schemas/CustomerViewModel'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardModel'

    CardModel:
      discriminator:
        propertyName: cardType
        mapping:
          creditcard: '#/components/schemas/CreditCardModel'
          debitcard:  '#/components/schemas/DebitCardModel'
      properties:
        cardNumber:
          type: string
        description:
          type: string
          example: "Huishouden"
        status:
          type: string
          pattern: (ACTIVE|BLOCKED|EXPIRED)
          description: current state of the card
        customerId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        cardType:
          type: string
          pattern: (creditcard|debitcard)
          example: creditcard
        expirationDate:
          type: string
          format: date
        issuer:
          description: Company issuing the card
          type: string
          example: Visa
      required:
        - cardNumber
        - cardType
        - status
        - expirationDate
        - issuer

    CreditCardModel:
      allOf:
        - $ref: '#/components/schemas/CardModel'
        - type: object
          description: CreditCard
          properties:
            cardNumber:
              type: string
              ## format: https://stackoverflow.com/questions/9315647/regex-credit-card-number-tests - holy cow ..
            cardHolder:
              description: Card holder
              type: string
              example: KRAMER
            securityCode:
              description: CVV
              type: string
              format: '[0-9]{3}'
              example: 656

    DebitCardModel:
      allOf:
        - $ref: '#/components/schemas/CardModel'
        - type: object
          description: DebitCard

    WithDrawModel:
      discriminator:
        propertyName: transfertype
        mapping:
          withdrawal: '#/components/schemas/WithDrawModel'
          transfer: '#/components/schemas/TransferModel'
      properties:
        from:
          $ref: '#/components/schemas/iban'
        amount:
          $ref: '#/components/schemas/MoneyModel'
        description:
          type: string
        transferType:
          type: string
          pattern: (withdrawal|transfer)
        authorizationMethod:
          type: string
          pattern: (creditcard|debitcard|authenticator|ideal)
        authorizationSignature:
          type: string
        transactionType:
          $ref: '#/components/schemas/transactionType'

    TransferModel:
      allOf:
        - $ref: '#/components/schemas/WithDrawModel'
      properties:
        to:
          $ref: '#/components/schemas/iban'

    TransactionModel:
      properties:
        timestamp:
          type: string
          format: date-time
        from:
          $ref: '#/components/schemas/iban'
        to:
          $ref: '#/components/schemas/iban'
        amount:
          $ref: '#/components/schemas/MoneyModel'
        description:
          type: string
        authorization:
          type: string
        transferType:
          type: string
          pattern: (withdrawal|transfer)
        authorizationMethod:
          type: string
          pattern: (creditcard|debitcard|authenticator|ideal)
        authorizationSignature:
          type: string
        transactionType:
          $ref: '#/components/schemas/transactionType'
      required:
        - from
        - to
        - amount
        - authorization

    TransactionViewModel:
      allOf: # Complete view
        - $ref: '#/components/schemas/TransactionModel'
      properties:
        id:
          type: integer
          format: int64

    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
        - code
        - message
      type: object